requirements.txt

dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
django-bootstrap3==9.1.0
django-multiselectfield==0.1.8
django-widget-tweaks==1.4.1
gunicorn==19.7.1
pytz==2017.3
static3==0.7.0
psycopg2>=2.6.1
